Optimize function and add check for incoming transaction
